BE-2072 - Move parseMonitorYAML test into yaml_utils_test
Address remaining CodeRabbit feedback: the test belongs alongside the
utility in yaml_utils_test.go rather than a new parallel test file.
